refactor(capabilities): :recycle: compose DejavuEasterEgg and CommandPalette easter eggs at feature level
Add ContentPage and ReadingContentPage feature wrappers that inject DejavuEasterEgg as
headerWrapper, making the easter egg wiring explicit at the feature composition boundary.
Move LayoutAdditionalContent from design-system/templates/ to features/ so its pwa,
matrix-rain, and easter-egg imports are legal under the architecture rules; inject
whiteRabbitEasterEgg and NeoRoomEasterEgg into CommandPalette here.
