+++ src/components/features/content-page/index.ts
@@ -1,0 +1,2 @@
+export { ContentPage } from "./content-page";
+export type { ContentPageProps } from "./content-page";
