+++ src/components/features/reading-content-page/index.ts
@@ -1,0 +1,2 @@
+export { ReadingContentPage } from "./reading-content-page";
+export type { ReadingContentPageProps } from "./reading-content-page";
